Move the kubeadm fuzzing into a separate package not imported by cmd/kubeadm
