+++ cmd/kubeadm/app/apis/kubeadm/install/install_test.go
@@ -20,9 +20,9 @@ import (
 	"testing"
 
 	apitesting "k8s.io/apimachinery/pkg/api/testing"
-	"k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm"
+	"k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/fuzzer"
 )
 
 func TestRoundTripTypes(t *testing.T) {
-	apitesting.RoundTripTestForAPIGroup(t, Install, kubeadm.KubeadmFuzzerFuncs(t))
+	apitesting.RoundTripTestForAPIGroup(t, Install, fuzzer.KubeadmFuzzerFuncs(t))
 }
